Lots of vision tuning
Got things to track reasonably well using these settings... but only if
the camera was attached to a separate laptop that ran GRIP and fed the
data into the RIO's networktable.  This code has the CameraServer stuff
commented out to try and avoid a conflict with the GRIP, but could be
re-enabled...or likely use the GRIP smartdashboard plugin instead.