Add shooterPID subsystem skeleton
Added the left and right side PID subsystems.  moved motors and encoders
to those classes, and then tweaked the original shooter class to access
the same objects through the new classes-- so no net functionality
change.  Next step is to implement the PID part and have the wrapper
"shooter" class handle sending the setpoint to both PID subsystems at
the same time (currently all the shooter functions take a motor
strength; this needs to change to be an RPM by default and add a
secondary direct motor strength for backup)