Updated shooter PID and integrated range
Altered shooterPID output to scale output of PID controller (in RPM) to
a motor speed, and then added equation to convert range into a desired
RPM.  This should give us enough code to spin up the shooter to a speed
appropriate for our range.  The initial equations for both are in
Shooter.java, and will need tweaking.