Shooter, vision, climbing, collector, and driveline changes before Unbag
Numerous changes to support upcoming robot debug:
1) DriverRight Trigger runs the RPM-based BasicShoot.  Use the
ShooterRPM widget in SmartDashboard ("SD") to change its speed.
Original motor-based function EngageShooter is still available in the
code but is unassigned.
2) Added command/SD button "AutoSmallPIDSequence" to invoke auto
distance calc plus use Nick's quick-and-dirty PID (called SmallPID in
the code) to spin the shooter up to the calculated RPM
3) Added cmd/SD button "AutoShootSequence" to do #2 using the full WPI
library.
4) Added cmd/SD button "AutoSmallPIDCenterAndShootSequence" to do #2 but
center the target using the vision PID first.
5) Added cmd/SD button "AutoCenterAndShootSequence" to do #3 but center
the target using the vision PID first.
6) Added SD text widget "TargetSetpoint" for changing the vision PID
setpoint for the target.  Defaults to zero, which was our original
assumption for "centered".  This allows user to compensate for the
off-center camera if that is needed.  Scale is -1 to +1.
7) Added turbo mode button (Driver JoyRight#2) to increase from default
75% speed to full 100% on driveline
8) Added remaining code to support collector deploy
9) Reordered PWMs to allow all "real" motor controllers to be plugged
into the ten available slots
10) To accomplish #9, we have to tie the new external collector to the
middle feederRoller PWM. Left a fake controller and functions in place
in case we ever have to separate the two later.
11) Added support for a camera light relay.
12) Added initial climb deploy and winch support.